Recover from thread exit in pre-main
The SEGGER JLink GDB server's RTOS integration will terminate the
primary thread when it loads the Zephyr RTOS threads as part of the
first thread info read. This causes the main thread to exit in an
asynchronous event while we're waiting for ThreadInfo in the break mode
event handler, which makes the thread lookup fail.

Attempt to recover from this by swapping to the first existing thread in
the thread cache and continuing execution.